Make precipitation checks work with probabilities.
Some weather integrations (For example,
https://www.home-assistant.io/integrations/nws/)
don't set a `precipitation` attribute in the
forecasts, but instead set a
`precipitation_probability` attribute. In these
cases, the integration will currently fail with a
`TypeError` trying to compare `None` to a `float`.
This fixes that problem by first checking whether
the forecast has a `precipitation` attribute and
falls back to `precipitation_probability` if
missing.